Add files via upload
A function to fit multiple light response curves. This script provides the same fit as the "Fitting_AQ_curves.R" script but makes it much easier. In that file, you're forced to modify the first few lines of a for loop. In this file, you only need to run the function code, then call the function with your data.