Error throwing update
While testing out some additional curves I had laying around it became clear that some data just will not fit the underlying model. This was a fatal problem because I had to code in the function for dealing with errors, meaning that none of the fits would be returned if one curve of many was problematic. 

So, I added in a tryCatch() call that allows the iteration to continue running through a problematic call. Now the function will return whatever the internal error message is when run. And, the data.frame of returned fits will contain NA's for any curve that caused issues. 

####### ####### ####### ####### 

Also. The function uses the slope of the relationship between Photo and PARi - for the first five points - as an initial starting value for quantum yield (phi). It occurred to me that not every curve thrown at this is necessarily in increasing order of PARi, and it is important that the slope estimate is taken at low light. To fix this potential problem I put in an internal ordering call that sorts each curve according to PARi.

####### ####### ####### ####### 

I added a few additional comments in the function.